Remove arctic from Qlib core to Contrib
