+++ qlib/data/__init__.py
@@ -15,7 +15,6 @@ from .data import (
     LocalCalendarProvider,
     LocalInstrumentProvider,
     LocalFeatureProvider,
-    ArcticFeatureProvider,
     LocalExpressionProvider,
     LocalDatasetProvider,
     ClientCalendarProvider,
