+++ setup.py
@@ -78,7 +78,6 @@ REQUIRED = [
     "dill",
     "dataclasses;python_version<'3.7'",
     "filelock",
-    "arctic",
 ]
 
 # Numpy include
